requirements.txt
# Core dependencies
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# AI/LLM integration
openai==1.3.5
anthropic==0.3.11

# CLI and utilities
click==8.1.7
python-dotenv==1.0.0
rich==13.6.0

# Data handling
pandas==2.1.3
sqlite3  # Built into Python

# Email processing
email-validator==2.1.0
beautifulsoup4==4.12.2
html2text==2020.1.16

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
black==23.10.1
flake8==6.1.0
pre-commit==3.5.0
